Factor out MouseHandler in separate class
