+++ MouseUnSnag.csproj
@@ -62,6 +62,7 @@
     <Compile Include="CommandLine\Options.cs" />
     <Compile Include="GeometryUtil.cs" />
     <Compile Include="Hooking\HookHandler.cs" />
+    <Compile Include="MouseHookHandler.cs" />
     <Compile Include="MyCustomApplicationContext.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="Program.cs" />
