+++ MyCustomApplicationContext.cs
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+using System;
 using System.Windows.Forms;
 using MouseUnSnag.CommandLine;
 using MouseUnSnag.Properties;
@@ -62,5 +63,6 @@ namespace MouseUnSnag
                 Visible = true
             };
         }
+
     }
 }
